Remove duplication for fileProperties()
