+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/SnapshotTransformInputsBuildOperationResult.java
@@ -25,14 +25,8 @@ import org.gradle.internal.hash.HashCode;
 import org.gradle.operations.dependencies.transforms.SnapshotTransformInputsBuildOperationType;
 import org.gradle.operations.execution.FilePropertyVisitor;
 
-import java.util.ArrayDeque;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Deque;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.TreeMap;
 
 public class SnapshotTransformInputsBuildOperationResult extends BaseSnapshotInputsBuildOperationResult implements SnapshotTransformInputsBuildOperationType.Result {
 
@@ -52,123 +46,16 @@ public class SnapshotTransformInputsBuildOperationResult extends BaseSnapshotInp
 
     @Override
     protected Map<String, Object> fileProperties() {
-        final Map<String, Object> fileProperties = new TreeMap<>();
-        visitInputFileProperties(new FilePropertyVisitor() {
-            Property property;
-            final Deque<DirEntry> dirStack = new ArrayDeque<>();
+        FilePropertyCollectingVisitor visitor = new FilePropertyCollectingVisitor();
+        visitInputFileProperties(visitor);
+        return visitor.getFileProperties();
+    }
 
-            class Property {
-                private final String hash;
-                private final Set<String> attributes;
-                private final List<Entry> roots = new ArrayList<>();
+    private static class FilePropertyCollectingVisitor extends BaseFilePropertyCollectingVisitor<FilePropertyVisitor.VisitState> implements FilePropertyVisitor {
 
-                public Property(String hash, Set<String> attributes) {
-                    this.hash = hash;
-                    this.attributes = attributes;
-                }
-
-                public String getHash() {
-                    return hash;
-                }
-
-                public Set<String> getAttributes() {
-                    return attributes;
-                }
-
-                public Collection<Entry> getRoots() {
-                    return roots;
-                }
-            }
-
-            abstract class Entry {
-                private final String path;
-
-                public Entry(String path) {
-                    this.path = path;
-                }
-
-                public String getPath() {
-                    return path;
-                }
-
-            }
-
-            class FileEntry extends Entry {
-                private final String hash;
-
-                FileEntry(String path, String hash) {
-                    super(path);
-                    this.hash = hash;
-                }
-
-                public String getHash() {
-                    return hash;
-                }
-            }
-
-            class DirEntry extends Entry {
-                private final List<Entry> children = new ArrayList<>();
-
-                DirEntry(String path) {
-                    super(path);
-                }
-
-                public Collection<Entry> getChildren() {
-                    return children;
-                }
-            }
-
-            @Override
-            public void preProperty(VisitState state) {
-                property = new Property(HashCode.fromBytes(state.getPropertyHashBytes()).toString(), state.getPropertyAttributes());
-                fileProperties.put(state.getPropertyName(), property);
-            }
-
-            @Override
-            public void preRoot(VisitState state) {
-
-            }
-
-            @Override
-            public void preDirectory(VisitState state) {
-                boolean isRoot = dirStack.isEmpty();
-                DirEntry dir = new DirEntry(isRoot ? state.getPath() : state.getName());
-                if (isRoot) {
-                    property.roots.add(dir);
-                } else {
-                    //noinspection ConstantConditions
-                    dirStack.peek().children.add(dir);
-                }
-                dirStack.push(dir);
-            }
-
-            @Override
-            public void file(VisitState state) {
-                boolean isRoot = dirStack.isEmpty();
-                FileEntry file = new FileEntry(isRoot ? state.getPath() : state.getName(), HashCode.fromBytes(state.getHashBytes()).toString());
-                if (isRoot) {
-                    property.roots.add(file);
-                } else {
-                    //noinspection ConstantConditions
-                    dirStack.peek().children.add(file);
-                }
-            }
-
-            @Override
-            public void postDirectory() {
-                dirStack.pop();
-            }
-
-            @Override
-            public void postRoot() {
-
-            }
-
-            @Override
-            public void postProperty() {
-
-            }
-        });
-        return fileProperties;
+        @Override
+        protected Property createProperty(FilePropertyVisitor.VisitState state) {
+            return new Property(HashCode.fromBytes(state.getPropertyHashBytes()).toString(), state.getPropertyAttributes());
+        }
     }
 }
